# nadzor-ai/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Инспектор ИИ — API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/ml/dataset/items": {
      "get": {
        "tags": [
          "ml"
        ],
        "summary": "Записи набора данных",
        "operationId": "dataset_items_api_v1_ml_dataset_items_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Status",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ml/dataset/items/{item_id}/curate": {
      "post": {
        "tags": [
          "ml"
        ],
        "summary": "Решение куратора по записи",
        "operationId": "curate_api_v1_ml_dataset_items__item_id__curate_post",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CurateInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/ml/dataset/versions/{version}/evaluate": {
      "get": {
        "tags": [
          "ml"
        ],
        "summary": "Метрики текущих машинных результатов на выборке выпуска",
        "description": "Precision / Recall / F1 / FPR по evidence_group (ТЗ 14.3).\n\nМашина «предсказала нарушение», если её статус находки был CANDIDATE.\nМетрики — по разделам и в целом, с размером выборки: без него точечная\nоценка на десятке записей выглядела бы весомее, чем есть.",
        "operationId": "evaluate_api_v1_ml_dataset_versions__version__evaluate_get",
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Version"
            }
          },
          {
            "name": "split",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(TRAIN|VALIDATION|HIDDEN_TEST)$",
              "default": "HIDDEN_TEST",
              "title": "Split"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ml/dataset/versions/{version}/export": {
      "get": {
        "tags": [
          "ml"
        ],
        "summary": "Выпуск набора в полях листа «Схема GOLD» (JSON или CSV)",
        "operationId": "export_version_api_v1_ml_dataset_versions__version__export_get",
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Version"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(json|csv)$",
              "default": "json",
              "title": "Format"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ml/evaluate": {
      "post": {
        "tags": [
          "ml"
        ],
        "summary": "Метрики по эталонной разметке (лист «Метрики», ТЗ 14.3)",
        "description": "Эталон читается только на время расчёта: он не сохраняется и не пишется в журнал.",
        "operationId": "evaluate_reference_api_v1_ml_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluateInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/ml/models": {
      "get": {
        "tags": [
          "ml"
        ],
        "summary": "Итерации дообучения",
        "operationId": "model_list_api_v1_ml_models_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ml"
        ],
        "summary": "Зарегистрировать результат обучения",
        "operationId": "register_model_api_v1_ml_models_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ml/models/{model_id}/approve": {
      "post": {
        "tags": [
          "ml"
        ],
        "summary": "Подписать публикацию модели",
        "operationId": "approve_model_api_v1_ml_models__model_id__approve_post",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ml/models/{model_id}/reject": {
      "post": {
        "tags": [
          "ml"
        ],
        "summary": "Отклонить модель",
        "operationId": "reject_model_api_v1_ml_models__model_id__reject_post",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ml/models/{model_id}/rollback": {
      "post": {
        "tags": [
          "ml"
        ],
        "summary": "Откатить опубликованную модель",
        "operationId": "rollback_model_api_v1_ml_models__model_id__rollback_post",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ml/report": {
      "get": {
        "tags": [
          "ml"
        ],
        "summary": "Отчёт по дообучению за период",
        "operationId": "report_api_v1_ml_report_get",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 366,
              "minimum": 1,
              "default": 7,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CurateInput": {
        "properties": {
          "approve": {
            "type": "boolean",
            "title": "Approve"
          }
        },
        "type": "object",
        "required": [
          "approve"
        ],
        "title": "CurateInput"
      },
...
      "EvaluateInput": {
        "properties": {
          "reference": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "items": {},
                "type": "array"
              }
            ],
            "title": "Reference"
          },
          "process_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Process Ids"
          }
        },
        "type": "object",
        "required": [
          "reference",
          "process_ids"
        ],
        "title": "EvaluateInput"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelInput": {
        "properties": {
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "dataset_version": {
            "type": "string",
            "title": "Dataset Version"
          },
          "precision": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Precision"
          },
          "recall": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Recall"
          },
          "f1": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "F1"
          },
          "false_positive_rate": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "False Positive Rate"
          },
          "per_category_metrics": {
            "additionalProperties": true,
            "type": "object",
            "title": "Per Category Metrics",
            "default": {}
          },
          "training_params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Training Params",
            "default": {}
          },
          "code_ref": {
            "type": "string",
            "title": "Code Ref",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "model_version",
          "dataset_version",
          "precision",
          "recall",
          "f1",
          "false_positive_rate"
        ],
        "title": "ModelInput"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}